+++ src/content/contentInjection.js
@@ -169,6 +169,7 @@ export const injectIntoPage = () => {
   });
 
   chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
+    console.log('update office links');
     if (request.action) {
       if (request.action === 'updateLinks') {
         updateOfficeLinks();